fix bugs when converting model to onnx

+++ centermask/modeling/fcos/fcos_outputs.py
@@ -388,8 +388,8 @@ class FCOSOutputs(object):
             box_cls = box_cls * ctrness[:, :, None]
         candidate_inds = box_cls > self.pre_nms_thresh
         # pre_nms_top_n = candidate_inds.view(N, -1).sum(1)
-        pre_nms_top_n = candidate_inds.reshape(N, -1).sum(1)
-        pre_nms_top_n = pre_nms_top_n.clamp(max=self.pre_nms_top_n)
+        pre_nms_top_n = candidate_inds.reshape(N, -1).sum(1).to(torch.float32)
+        pre_nms_top_n = pre_nms_top_n.clamp(max=self.pre_nms_top_n).to(torch.int64)
 
         if not self.thresh_with_ctr:
             box_cls = box_cls * ctrness[:, :, None]
